Fix nix flake build issues
This will use the deps.nix file from the commit that is built.

It also makes sure that the build can run in pure evaluation mode.
To make this possible, it only fetches jetzig from github if v14 is built.
Otherwise it just builds from the current directory.

Note that builds with both zig 0.15.1 as well as master currently fail,
so only the v14 build works.